+++ api/account_analytic_tag.go
@@ -12,11 +12,11 @@ func NewAccountAnalyticTagService(c *Client) *AccountAnalyticTagService {
 	return &AccountAnalyticTagService{client: c}
 }
 
-func (svc *AccountAnalyticTagService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountAnalyticTagService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountAnalyticTagModel, name)
 }
 
-func (svc *AccountAnalyticTagService) GetByIds(ids []int) (*types.AccountAnalyticTags, error) {
+func (svc *AccountAnalyticTagService) GetByIds(ids []int64) (*types.AccountAnalyticTags, error) {
 	a := &types.AccountAnalyticTags{}
 	return a, svc.client.getByIds(types.AccountAnalyticTagModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountAnalyticTagService) GetAll() (*types.AccountAnalyticTags, erro
 	return a, svc.client.getAll(types.AccountAnalyticTagModel, a)
 }
 
-func (svc *AccountAnalyticTagService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountAnalyticTagService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountAnalyticTagModel, fields, relations)
 }
 
-func (svc *AccountAnalyticTagService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountAnalyticTagService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountAnalyticTagModel, ids, fields, relations)
 }
 
-func (svc *AccountAnalyticTagService) Delete(ids []int) error {
+func (svc *AccountAnalyticTagService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountAnalyticTagModel, ids)
 }
