+++ api/account_bank_statement.go
@@ -12,11 +12,11 @@ func NewAccountBankStatementService(c *Client) *AccountBankStatementService {
 	return &AccountBankStatementService{client: c}
 }
 
-func (svc *AccountBankStatementService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountBankStatementService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountBankStatementModel, name)
 }
 
-func (svc *AccountBankStatementService) GetByIds(ids []int) (*types.AccountBankStatements, error) {
+func (svc *AccountBankStatementService) GetByIds(ids []int64) (*types.AccountBankStatements, error) {
 	a := &types.AccountBankStatements{}
 	return a, svc.client.getByIds(types.AccountBankStatementModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountBankStatementService) GetAll() (*types.AccountBankStatements, 
 	return a, svc.client.getAll(types.AccountBankStatementModel, a)
 }
 
-func (svc *AccountBankStatementService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountBankStatementService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountBankStatementModel, fields, relations)
 }
 
-func (svc *AccountBankStatementService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountBankStatementService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountBankStatementModel, ids, fields, relations)
 }
 
-func (svc *AccountBankStatementService) Delete(ids []int) error {
+func (svc *AccountBankStatementService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountBankStatementModel, ids)
 }
