+++ api/account_cashbox_line.go
@@ -12,11 +12,11 @@ func NewAccountCashboxLineService(c *Client) *AccountCashboxLineService {
 	return &AccountCashboxLineService{client: c}
 }
 
-func (svc *AccountCashboxLineService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountCashboxLineService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountCashboxLineModel, name)
 }
 
-func (svc *AccountCashboxLineService) GetByIds(ids []int) (*types.AccountCashboxLines, error) {
+func (svc *AccountCashboxLineService) GetByIds(ids []int64) (*types.AccountCashboxLines, error) {
 	a := &types.AccountCashboxLines{}
 	return a, svc.client.getByIds(types.AccountCashboxLineModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountCashboxLineService) GetAll() (*types.AccountCashboxLines, erro
 	return a, svc.client.getAll(types.AccountCashboxLineModel, a)
 }
 
-func (svc *AccountCashboxLineService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountCashboxLineService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountCashboxLineModel, fields, relations)
 }
 
-func (svc *AccountCashboxLineService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountCashboxLineService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountCashboxLineModel, ids, fields, relations)
 }
 
-func (svc *AccountCashboxLineService) Delete(ids []int) error {
+func (svc *AccountCashboxLineService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountCashboxLineModel, ids)
 }
