+++ api/account_config_settings.go
@@ -12,11 +12,11 @@ func NewAccountConfigSettingsService(c *Client) *AccountConfigSettingsService {
 	return &AccountConfigSettingsService{client: c}
 }
 
-func (svc *AccountConfigSettingsService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountConfigSettingsService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountConfigSettingsModel, name)
 }
 
-func (svc *AccountConfigSettingsService) GetByIds(ids []int) (*types.AccountConfigSettingss, error) {
+func (svc *AccountConfigSettingsService) GetByIds(ids []int64) (*types.AccountConfigSettingss, error) {
 	a := &types.AccountConfigSettingss{}
 	return a, svc.client.getByIds(types.AccountConfigSettingsModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountConfigSettingsService) GetAll() (*types.AccountConfigSettingss
 	return a, svc.client.getAll(types.AccountConfigSettingsModel, a)
 }
 
-func (svc *AccountConfigSettingsService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountConfigSettingsService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountConfigSettingsModel, fields, relations)
 }
 
-func (svc *AccountConfigSettingsService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountConfigSettingsService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountConfigSettingsModel, ids, fields, relations)
 }
 
-func (svc *AccountConfigSettingsService) Delete(ids []int) error {
+func (svc *AccountConfigSettingsService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountConfigSettingsModel, ids)
 }
