+++ api/account_invoice_cancel.go
@@ -12,11 +12,11 @@ func NewAccountInvoiceCancelService(c *Client) *AccountInvoiceCancelService {
 	return &AccountInvoiceCancelService{client: c}
 }
 
-func (svc *AccountInvoiceCancelService) GetIdsByName(name string) ([]int, error) {
+func (svc *AccountInvoiceCancelService) GetIdsByName(name string) ([]int64, error) {
 	return svc.client.getIdsByName(types.AccountInvoiceCancelModel, name)
 }
 
-func (svc *AccountInvoiceCancelService) GetByIds(ids []int) (*types.AccountInvoiceCancels, error) {
+func (svc *AccountInvoiceCancelService) GetByIds(ids []int64) (*types.AccountInvoiceCancels, error) {
 	a := &types.AccountInvoiceCancels{}
 	return a, svc.client.getByIds(types.AccountInvoiceCancelModel, ids, a)
 }
@@ -36,14 +36,14 @@ func (svc *AccountInvoiceCancelService) GetAll() (*types.AccountInvoiceCancels, 
 	return a, svc.client.getAll(types.AccountInvoiceCancelModel, a)
 }
 
-func (svc *AccountInvoiceCancelService) Create(fields map[string]interface{}, relations *types.Relations) (int, error) {
+func (svc *AccountInvoiceCancelService) Create(fields map[string]interface{}, relations *types.Relations) (int64, error) {
 	return svc.client.create(types.AccountInvoiceCancelModel, fields, relations)
 }
 
-func (svc *AccountInvoiceCancelService) Update(ids []int, fields map[string]interface{}, relations *types.Relations) error {
+func (svc *AccountInvoiceCancelService) Update(ids []int64, fields map[string]interface{}, relations *types.Relations) error {
 	return svc.client.update(types.AccountInvoiceCancelModel, ids, fields, relations)
 }
 
-func (svc *AccountInvoiceCancelService) Delete(ids []int) error {
+func (svc *AccountInvoiceCancelService) Delete(ids []int64) error {
 	return svc.client.delete(types.AccountInvoiceCancelModel, ids)
 }
